Stored Item audit for pallet

+++ database/migrations/2025_02_20_084056_add_scopes_to_stored_item_audits_table.php
@@ -15,7 +15,7 @@ return new class () extends Migration {
     {
         Schema::table('stored_item_audits', function (Blueprint $table) {
             $table->string('scope_type')->index()->comment('Fulfilment|Pallet')->nullable();
-            $table->unsignedSmallInteger('scope_id')->nullable();
+            $table->unsignedInteger('scope_id')->nullable();
             $table->index(['scope_type', 'scope_id']);
         });
     }
